+++ tjukica_zadaca_3/tjukica_zadaca_3/Composite/OrgJedinica.cs
@@ -98,9 +98,9 @@ namespace tjukica_zadaca_1.Composite
             return retVal;
         }
 
-        public override double DajZaradu(TipVozila tipVozila, DateTime datum1, DateTime datum2)
+        public override decimal DajZaradu(TipVozila tipVozila, DateTime datum1, DateTime datum2)
         {
-            double retVal = 0;
+            decimal retVal = 0;
             foreach (var item in tvrtkaComponents)
             {
                 retVal += item.DajZaradu(tipVozila, datum1, datum2);
